refactor: decouple bookings handler from API v2 with projection-based repository
- Create GetBookingsRepository with findManyForWeb() and findManyForApiV2() methods
- findManyForWeb() returns minimal fields for web UI (excludes sensitive fields like credentialId, meetingPassword)
- findManyForApiV2() returns all fields for backward compatibility with API v2
- Both methods share the same filtering logic, differing only in SELECT statements
- Update tRPC handler to use findManyForWeb() for optimized web queries
- Update getAllUserBookings to use findManyForApiV2() for API v2 compatibility
- Add DI module and container for the new repository
