+++ packages/features/bookings/di/GetBookingsRepository.module.ts
@@ -1,0 +1,25 @@
+import { GetBookingsRepository } from "@calcom/features/bookings/repositories/GetBookingsRepository";
+import { bindModuleToClassOnToken, createModule, type ModuleLoader } from "@calcom/features/di/di";
+import { moduleLoader as prismaModuleLoader } from "@calcom/features/di/modules/Prisma";
+import { DI_TOKENS } from "@calcom/features/di/tokens";
+
+const thisModule = createModule();
+const token = DI_TOKENS.GET_BOOKINGS_REPOSITORY;
+const moduleToken = DI_TOKENS.GET_BOOKINGS_REPOSITORY_MODULE;
+
+const loadModule = bindModuleToClassOnToken({
+  module: thisModule,
+  moduleToken,
+  token,
+  classs: GetBookingsRepository,
+  depsMap: {
+    prismaClient: prismaModuleLoader,
+  },
+});
+
+export const moduleLoader = {
+  token,
+  loadModule,
+} satisfies ModuleLoader;
+
+export type { GetBookingsRepository };
